Derive lemma language from lexeme language for new lexemes.
On Special:NewLexeme show the lemma language field only if the item
selected for the lexeme language does not already contain a language
code.

The language is determined by checking the selected item for a statement
with the property ID defined in the LexemeLanguageCodePropertyId config
variable. This only gets the value of the first statement returned by
the API and does not consider ranks.

In order to reuse the validation of the HTMLContentLanguageField it
currently expects the value of the language code statement to be a two
letter language code which is element of MediaWikiContentLanguages.

Bug: T160518
Change-Id: I7d33e7a7bebb6b62666dc5ddf73d1acc20f006f1