Fix and properly document contract of Lexeme::isEmpty
Think of "isEmpty" as the oposite of "can be emptied". Is the entity
object already empty, or is there anything that still can be transfered
to an other entity and then removed from the original one? This is why
this method exists: to make sure an entity is not merged or turned into
a redirect before transfering all transferable content to the target.

Change-Id: I42c2e15338239b1534bf8456ac91d7384fe77e70