Remove unused interfaces
These do not provide an actual benefit. Instead we can simply type
hint against Lexeme. All code that will ever work with these aspects of
a Lexeme is in the same code base anyway. So these interfaces are not
needed for dependency inversion. It is also very unlikely we will ever
reuse any of these Lexeme specific aspects in an other entity type.

Let's remove this for now to make working on this code easier, and
possibly add it later when we really need it.

Change-Id: Ibe1defaa693220e4882774076afe1175c08f2237