Use retrieving label lookup form lexical category and language labels
Warning! This is more a band-aid patch. For a proper solution we need
to "teach" the existing EntityParserOutputDataUpdater how to deal with
Lexemes, or better: make it possible to override the default
EntityParserOutputDataUpdater with one that knows about Lexeme-specific
concepts.

We need to do this anyway the moment we have Statements inside Forms
and Senses. I would tackle the issue then, and call T162698 "done" with
this patch.

Bug: T162698
Change-Id: Iafd868db2fe4e8408f59439b9af2344e60d37bb3