Also check ID in Lexeme::isSufficientlyInitialized
The ID also is a field that can never be set back to null. All three
fields can be initialized once, but this decision can never be
reverted.

This does have zero impact on the code using this, because the ID is
already checked. But it makes the Lexeme implementation more consistent.

Change-Id: I3136d6cc8b09227bbc9c306ab83a2ffec0c5416b