Enable editing of forms in the form list view.
This adds a new ControllerViewFactory in order to wire together the
lexemeview, lexemeformlistview and lexemeformview including the toolbar
and ToolbarViewController for lexemeformview which are needed for
editing.

It also adds an intial LexemeForm model for the datamodel frontend
which at the moment does not inherit from Entity. It is only needed to
provide the equals-method for the ToolbarControllerView.

Persistence is not implemented yet. It uses a fake model instead of an
EntityChanger that always resolves the "save" promise.

Bug: T162339
Bug: T162340
Bug: T162341
Change-Id: Ie36d5e5dbd80706f26f91d57abe742decfad3c3d