Add package-lock.json
Since version 5, npm creates this file to improve dependency management. The
file is intended to be commited to source control [1]. It should not interfere
on system where an older version of npm is used (ahem, jenkins...). I recommend
all developers to upgrade to version 5 because it's much more stable and a lot
faster.

[1] https://docs.npmjs.com/files/package-lock.json

Change-Id: Icc91a224e264c9b33df70254958b0af95a463d03