Report errors to the UI
To report errors we utilize default core's error reporting mechanism,
so that API users will not need to learn about custom error format.

To display errors in the UI correctly we reformat API error response
to the format that our UI "framework" understands, so it can be
migrated later without changing API code and response structure.

Change-Id: I5771ad0446074cc98ad22fe172ba286550c3bab4