Make Lexeme constructor accept FormSet only
This is mostly for consistency. The class exists. There is the NewForm
builder which should accept other stuff for convenience. Here we want to
be strict.

Next step is to make all the getters return FormSet objects too. But
this touches much more code and will be done later, see I1af251f.

I'm also fixing a few mistakes in related code.

Bug: T164742
Change-Id: I5385781b0109b4fe0023c27fa987cc1978ffc525