Remove meaningless @group WikibaseLexeme from all tests
Look up the term "cargo cult programming". In my opinion this here is a
good example for that.

The basic idea if test groups is to be able to quickly and conveniently
run a specific subset of tests. But what subset defines a
"WikibaseLexeme" group? It's literally the entirety of the WikibaseLexeme
extension. This tag would need to be added to every single test. But why
should we do this, and take care of it? Note that it is already
inconsistently used (in most tests, but not in all of them). This was
obviously never a concenious decision. It was forgotten in a few cases,
but that never caused any issues anywhere – except somebody wants to use
it some day. It will not do what that person expects then.

If somebody wants to run the WikibaseLexeme tests, he can simply do this
based on the directory. Done.

Change-Id: I151eb3e04857fe6fb35a16c6538a005fb78c34ab