Add a script for updating outdated serialized data in the DB (dev use only)
The script adds nextFormId to the serialized lexeme data for
entries that don't have it.
For possible future changes in the lexeme serialization format
(before the production roll out), this script could be handling
those changes as well.

Script also adds forms and senses fields to even older
lexeme data.

Bug: T178045
Change-Id: I41f1bd1162d8bb2ef6c485a825893436611c5163