Fix minor type violation in LexemeHandler "history" action override
We know this does not make a difference because there are only two
classes implementing the Page interface. But the code does not know.
I get a warning that Article::newFromWikiPage can only be called with
a WikiPage, which is true and what the if() above should check for.

See I55edb34 as well as Icf6055f for reference.

Change-Id: I888e24b91c50d7f926090b7839fae39af6e5caad