Make FormId extend EntityId
This patch is intentionally as small as it can be. The FormId class is
now an EntityId, but it does not allow prefixed ("foreign") IDs. This is
currently not needed. I'm aware this violates LSP, but I suggest to accept
this for now and change it later when we need FormIds that point to foreign
Wikibase repositories.

Bug: T180469
Change-Id: I505e5dd96d665196885a019c73754aea350b1ac0