Fix violation of using package private serializer implementations
These two implementations are marked as "package private", which means
it is actually forbidden to instantiate or use them in any way that
sidesteps the public methods provided by the serializer factory.

This is not some personal opinion or anything I came up with, but
clearly marked in the contract of the Wikibase DataModel Serialization
component.

Bug: T178994
Change-Id: I4cb1591c626c4b0a8b6b32a4dc1da7244baa3c5a