Remove not needed inline @var type hints from LexemePatcher
The type of these variables is already specified via the instanceof
directly above. I forgot to remove these @var in the patch that
originally introduced the "case … instanceof" style, see I0db1e31.

You might wonder why I'm moving the "@var Form" so far up. The answer
is that the individual cases in a switch do not form independend
scopes in PHP. Variables introduced in one scope "bleed" into the next.
This is all fine here, as the $form variable never changes type or
meaning. I hope this becomes a little more obvious with the single
@var above.

Change-Id: I0b7c29dd7ea823745c7d20491b79bdb33200f54f