Move @var type hints in LexemeDiffer/Patcher up to the @param header
This patch moves inline @var type hints further up to the method headers.
The benefits of this are:
* No extra @var are needed.
* The expected types are more visible, right away in the expected place
  in the method documentations, and not only inline or in @throws
  documentations.

Note that even if the method headers still type hint against
EntityDocument (they need to, to conform to the interfaces the methods
implement), calling these implementations with anything but a Lexeme is
an actual mistake, and imediatelly causes a LogicException (which
basically means "you are doing it wrong").

Change-Id: I99e61c5e2740351050eb19aa83019c61eb70e1cd