Set language code in Form edit summaries
This language code field is currently unused in the messages. However,
I figured it does not hurt and might be helpful in the future to have
the language code in the field that is reserved for the language code.
For the summaries touched here this is only possible if there is only
one language code.

Change-Id: I328712d8cb70b9d0dc86850bffc498f1c687182f