Remove convenience function Lexeme::hasForm
Lexeme::hasForm is a convenience function that is barely used
(only one or two times, depending on the branch), and already
has two equally valid alternatives:
* $lexeme->getForm( $formId ) !== null
* $lexeme->getForms()->hasForm( $formId )

By removing it we reduce the overall complexity of the already
quite big (400+ lines) Lexeme class.

This patch got motivated by the dicsussion in Ic3356de.

Change-Id: I96b7f528b165fc51e3a307066ab6114d3b3acf3f