Disable Senses, the softer way
This uses a config variable to make it easier to develope senses
also having better control over API as well.

It should be better if we had them as settings and inject it properly
but given the interim nature of the whole thing, I went with simpler solution
also another reason is that I'm lazy.

Bug: T187196
Change-Id: Id16a972a8ff6357f72b2cb1d90bccb7c427115d4