Add FormDifferTest (unit test)
This patch also throws exceptions in our Diff / DiffOp classes
to do with Forms that are missleading.

For example, if a ChangeFormDiffOp has diff ops the default
implementaiton of getOperations and getArrayCopy would both
return empty arrays.

As with Idd0781816803dbb49a3e490696109c0addb0e656 lets not secretly
return bad or miss leading data if these things are called, instead
fail hard and fast so that we know.

Change-Id: I02011398d5a79a532e569f4f049b4f6df2104685