[cirrus] Fix forms search
- use a match query to properly use the analyzer on the form id field
- fallback to 'und' when the language code is unknown

we could consider changing the mapping for form ids but if we do so a
match query will still be required for the transition.
I don't think there can be ambiguities on the ids because we do
casefolding. But if we believe it may happen another patch is required
to change the mapping.

Bug: T195453
Change-Id: I41ac46fc4d38c2b86042be36b21327b7f4844c61