Add @covers tags for Add/Change/RemoveFormDiff to FormDifferTest
These three classes are currently reported as one of the biggest project
risks, according to the calculated CRAP factor.

https://doc.wikimedia.org/cover-extensions/WikibaseLexeme/dashboard.html

This does not necessarily mean much. It's just a number that appears so
high because these classes have so many tiny methods. But it appears this
is easy to fix, with little investment.

I did a project wide search for constructor calls to these classes (in
non-test code), and there is only a single one in the FormDiffer class.
This essentially means these three classes can be considered part of
FormDiffer and part of FormDifferTest.

Change-Id: I5da6d03c8e78c58b52ebfe7d07c37e08406037b9