Only register API modules and other repo-specific stuff in repo-only mode
Wikis running only "client" part of Wikibase must be able to
handle e.g. items having lexeme references in their statements.
The way to achieve that could be having the Lexeme enabled on
those wikis. But currently WikibaseLexeme has number of repo-specific
functionality, and simply enabling the extension there would break
things.

It seems the way to go would be to have a common, core "Lexeme"
functionality in some library, and have repo-, and client-specific
stuff in separate MediaWiki extensions.

Achieving this is bound to take time, so as a temporary measure
this moves repo-specific things out of extension.json to a
registration method that only registers API modules, special pages,
resource loader modules and so on when the Wikibase Repo component
is enabled

TODO: how to tests all this?

Bug: T195615
Change-Id: I9cc3b7f0892353523b9061f8debddaf92bd026dc