Add nextSenseId to Lexeme
Similar to the nextFormId, a Lexeme gains a nextSenseId, which holds the
numeric ID to use for the next sense of this lexeme, to ensure that
sense IDs are not reused when senses of a lexeme are removed and added.
The serialization layer learns to serialize and deserialize this new
field, but a missing nextSenseId is accepted as long as the senses are
empty. Since storing senses is not yet supported, that should be enough
to provide for a smooth upgrade on all wikis, unless someone has
manually inserted lexeme serializations with senses into page storage.

Some lines of code, especially the added TODOs, anticipate the future
addition of some kind of SenseSet class similar to the existing FormSet,
which would have a method to return the maximum sense ID used in the
set.

Bug: T198194
Change-Id: I4401096d37e718fc4bbca8d9685ce3e22ca29dd7