Always construct SenseId including lexeme ID
As preparation for making SenseId a subclass of EntityId, make sure that
we always instantiate it the way it should be instantiated: with the
full sense ID, including the lexeme ID part (“L1-S1”), just like for
form IDs (“L1-F1”). The NewSense factory helper is modified to always
add the lexeme ID (currently hard-coded “L1”), and should itself be
called with just the sense-specific part.

Bug: T199097
Change-Id: I2ead9079eed0b40aedaaa340267610f1997ac224