Selenium: add selenium-daily NPM script
The script is needed to run the new Docker-based Jenkins job that runs daily and targets beta cluster.
selenium-test script and NPM packages are dependencies.

The patch fixes some references to page objects in core. There are still
some references to page objects in core that could not be fixed becase
the pages are not yet in wdio-mediawiki NPM package. Additionally, tests
reference files from Wikidata. That is also not fixed.

Change-Id: I16b97b5a91e57214506c78726250c0457b4c0a5e
Job: WikibaseLexeme-npm-browser-run-selenium-daily-node-6-docker
Bug: T188742
Bug: T194252