Prevent empty form representations in front-end
Do not allow the user to remove the last form representation by
disabling the “remove” button if there is only one representation.
Styles are slightly updated so that only enabled buttons look active.

The color for the disabled button is Base30 from the Wikimedia Style
Guide, described as appropriate for “disabled elements text”.

Bug: T199703
Change-Id: Ie8ce68013cff5b91cfaaa9d7efc8ad7c3c383156