Correctly serialize glosses in SenseSerializer
Now that a sense’s glosses are a TermMap, not a POJO (see commit
c53d7100d9, change I9948938309), we have to serialize them properly
instead of storing them in the serialization as-is, just like in
FormSerializer for the representations. (In fact, the duplicate code
should be refactored into some kind of TermMapSerializer in the future.)

This also registers the SenseSerializer tests with MediaWiki, so that
next time the tests break due to a change like this, we hopefully
notice it earlier.

Bug: T198197
Change-Id: I97971f85677f2c66ff313488cec8a13d03f20940