Store real form/sense ID in BlankForm/BlankSense
When a BlankForm or BlankSense has been used to create a real Form or
Sense, remember its entity ID and return that on subsequent requests for
the BlankForm/BlankSense’s entity ID. This makes sense because a
BlankForm/BlankSense is not typically used to create more than one Form
or Sense, and is necessary because not all references to the
BlankForm/BlankSense are replaced with the real Form or Sense when they
should be.

The practical effect of this is that the wbeditentity API will now
return the ID of newly created forms and senses, instead of “null”. A
test for this behavior is also added.

Bug: T200255
Change-Id: I819783c5d39f7b57912b170a57b88e8d2b9a085a