Set input to sense ID in Sense expert
Normally, the input for an expert is initialized with the text rendering
of the value, e. g. the label of an item. However, if we keep this
behavior for senses, then it’s difficult to edit a sense, because what
you get in the input (lemmas and gloss) is totally unrelated to what
needs to be in the input to select this or any other sense (sense ID).
Instead, we initialize the input with the plain entity ID. (We might be
able to remove this later with improved search support.)

Bug: T196269
Change-Id: Ie72a410b367141f9a6fafec5961c71bd96efc9f8