Actually format a summary in SpecialNewLexemeTest
Before this change the summary formatter methods are not mocked,
and so null would be the result of the format call.
This would then be passed all the way down to the DB layer for
storage which would fail.

I strognly feel that this should have been picked up by Wikibase
CI, as Wikibase now runs with Lexeme and the patch that likely caused
this was only merged yesterday?

Bug: T202114
Change-Id: I1297903a9d37cf3b969d34abaf9610d3d91f65de