Add language fallback to SenseIdHtmlFormatter
This adds language fallback support to the HTML formatter for sense
values.

This removes the use of the text formatter as part of the html formatter
as the interface no longer provides us with enough infomation (it only
provides us with a text string but we need a text string and the
language).

Bug: T200983
Change-Id: I18c0a42a2a90531e8f4b02b045a7ce64b5fcb6e0