LexemeMergerTest: avoid conflicting lemmas
NewLexeme randomly generates lemma language and term if not passed
a custom lemma. There is a small but existing chance that a source and
target lexeme are generated that have identical languages, which then
almost certainly have contradicting terms - a condition that is
validated against and hence should not be occuring in the default test
data meant to act as a basis for further refinement.

Bug: T203149
Change-Id: I96ee5cf77bff2745de3309b1829fd2ad65b49375