Use hyphenation for long Lexeme Lemmas
Allow browers to apply auto-hyphenation to long lemmas in the lexeme
header.
This also ensures that the buttons (e.g. edit) in the lexeme header,
which show little icons before the call-to-action, are not wrapped when
there is only little space to render them.
Additionally it ensures that the fallback word-wrap can come to effect
on children of our flex elements:
https://css-tricks.com/flexbox-truncated-text/

Bug: T195367
Change-Id: Id3cee249599aefeb2dcbd1d50be35190b64aff61