LexemeEnableSenses true by default.
As we can't change the config loaded in CI this is needed to be able to
run browser tests for Senses.

I have checked in WMF production and the setting is already set to
false in the configs there so this will be a noop for our wikis.

This is ultimately needed in order to merge Ifa48f76 with its tests.

Change-Id: I079549c83f41812cc496202dbfbb876a9e10c31b