Undeprecate FormDiffer::diff
For all users inside of WikibaseLexeme this is actually the more sensible method to use.
It is perfectly fine to have both this normal-style method and the somewhat-weird
diffEntities one that acts as adapter for EntityDifferStrategy.

The generic approach of EntityDifferStrategy itself is also questionable as per
https://www.entropywins.wtf/blog/2017/05/24/generic-entity-handling-code/
but that is a different issue.

Change-Id: I92ba0d2c2692bf9b85eb84428c364c3844443d60