Clone source entity in ChangeOp{Form,Sense}Clone
ChangeOpFormClone and ChangeOpSenseClone reset statement GUIDs when
copying over statements to the target entity, which results in bugs when
those statements are still the same objects as in the original source
entity (it means that we save a version of the source entity with
statement GUIDs erased when merging lexemes). To avoid this, copy the
source entity in the constructor.

Bug: T201605
Change-Id: Ief59583f3af25a6cc764b3d5705f4ce1f043f3e1