Form: do not validate at construction time
Validating forms on creation complicates using ChangeOps which build on
the idea of adding information step-by-step.
This was worked around so far by using BlankForm but now is further
complicated by the need to use a Form including valid FormId inside a
ChangeOp generating statement GUIDs which creates a hen and egg
problem.

Bug: T194732
Change-Id: Ibaa9f01eea62c7ef29060b324ff7d9f2e904566b