form statement groups: ids without lexeme part
The statement group ids (anchors) of subentities should be consistent
with the anchors we introduced for the subentities themself, that is
without the lexeme part.
Resolves the full ids introduced in I559609e8bc0f65ba0f41282bb3447ca80aa5bb6d

This also resolves the problem that previously, after the creation of a
new form, the resulting statatement group ids were not prefixed
correctly - because the form id was unknown at the time of creation.

Bug: T196226
Change-Id: Icd04008b2544ddaf18879fc1945dc7519f7513a6