Remove accidental complexity
Instead of injecting a LexemeValidatorFactory into ChangeOpDeserializers
and having them retrieve the validators they actually need from it,
directly inject those validators. This allows us to completely remove
LexemeValidatorFactory, which removes a lot of code, especially in the
unit tests.

Note that the validators which WikibaseLexeme.entitytypes.repo.php sets
up are empty and do not validate anything, which means that the API lets
you create lexemes whose language or lexical category items do not
exist. (Special:NewLexeme prevents this.)

Change-Id: I4662433912db342bbb9a773af6adedb4c825209c