Rename bot flag parameters and properties
This boolean does *not* indicate whether the user “is” a bot, in the
sense of being in a certain group or having a certain right. It is a
request parameter which can be set freely; if the user has the 'bot'
*right* (conventionally by virtue of being in a group called “bot”, but
other groups may also include this right), and this flag is set, then
the edit is marked as a bot edit, otherwise the flag is ignored.

https://www.mediawiki.org/wiki/Manual:Bots#Bot_right,_group,_and_flag

Change-Id: Ib1cdd66411234997e7c0718720366829f456e31b