Ensure lexemeview is loaded during debugging
This fixes a bug that the js debugging with ?debug=true
query param was broken on Lexeme pages.
The root cause was a race condition between the Wikibase code
initializing the entity view and the WikibaseLexeme code initialising
our ViewFactoryFactory and our $.wikibase.lexemeview.

This resolves that race condition by having the Wikibase code wait for
our promise to be resolved.

Bug: T204039
Depends-On: I6d7236da866c79131b58b6b2a570cfd48d3dd610
Change-Id: I00be0a92df2d9abb957b528622072961875af3bf