Update upstream wikimedia eslint config
This updates the wikimedia eslint config because we want the rules that
ensure that our JS doesn't break on IE11. However, that new release also
includes rules that forbid a lot of jquery syntax. This rules are
reasonable, but beyond this commit or ticket. There will be a new ticket
adjusting our code for these rules.

Bug: T209932
Change-Id: I769e5cecac2f1a6bc9fabca95ef9a6e2ee8b7fb0