Check if user blocked on Special:MergeLexemes Load
If user is blocked show an error right away on Special:MergeLexemes
not just when the users submits the page.

This is better for the user and makes testing easier for us.

Lexeme version of: I6258f9824867936601dbbff0b53f7b109f9474e8

Also add browser tests to cover this

Bug: T211120
Depends-On: Icf94458eb9c9a1fa25e5ef9291d9fc2c54a5567c
Change-Id: I118a14930edd2faef20709ba7a65e6baf278aa45