Suppress compatibility errors only in tests
Since the tests are intended to be run with modern-ish browsers or node,
compatibility is less of an issue there and we can suppress errors
unless they become a problem. However, in the live code we want eslint
to complain about syntax that might be a problem in IE11 and similar
browsers.

Bug: T210744
Change-Id: Ib7ec340fdb56bd96ac4bdc54c007a6766388780d