Remove bogus "0.0.0" copy-pasted from BoilerPlate extension
Motivated by the discussion in T213066. I'm sure this is a mistake.
Openly saying that this codebase does have "version 0.0.0" does not
make much sense, in my opinion. I believe this is a copy-paste mistake
either from the BoilerPlate extension, or from package.json that seems
to require this dummy version number. extension.json does not. It can
be omited there.

Change-Id: I9d48e70920be6137da6b659e70b957a9616e18cf