Bug, Make Lexeme::equals actually check senses & test
This fixes T211438...
This also is the reason the test for undoing in the previous patch
actually passes.

This also identifies a difference between sense merging tests
and what is actually happening.
The failing sense merge tests have been commented out and replaced
with tests for the current behaviour.
We can follow up and check what the desired behaviour actually is.

Bug: T211438
Change-Id: I2c6eab2ee0fd61e18f5818641b3f81a7f3e4a9ac