Cleanup commented out tests in LexemeSensesMergerTest
The BDD for merging senses is described in:
https://phabricator.wikimedia.org/T201605

This states that:
"AND each individual Sense including Glosses and statements is added as
a new Sense to the target Lexeme"

So the commented out tests are just wrong (probably a left over from
copy pasta?)
So this patch removes those tests.

This patch also cleans up the names of some of the test cases making it
clearer what is being tested there merging -> copying when it comes to
senses most of the time.

This is a followup to I2c6eab2ee0fd61e18f5818641b3f81a7f3e4a9ac

Bug: T211438
Change-Id: I6bf7566aecaa1f1f555cbb31eef59ce10183517f