Use default overlay for item selector widget
The popup can be very long and in viewports with a low height, it may
popup towards the top. However, then it might appear behind pagename
and parts of the usermenu. The cause is that the page content
constitutes its own stacking context for security reasons.

This patch remedies that problem using the option to place the popup in
a container outside of the body.

Bug: T199409
Change-Id: I0e20135a8647061082860273f846621e88545cad
Depends-On: I55f4d0334403b5b9769a3f79eed147c6eed84c53