Give elastic on beta more time to recognize created items
Elastic doesn't seem to pick the items up quick enough on beta and so
adding them as the new language item fails.

The change to startHeaderEditMode is necessary because, for some unknown
reason, the lexeme language and lexical category are loaded
asynchronously after the header edit-button has been clicked. We have to
wait for them to be loaded before we add our own values.

Bug: T222449
Change-Id: I21b588651a58d77ec4a19f4acfbda40bb8bc36d6