Retry inputs in Lexeme header browser tests
This is supposed to fix the problem that a newly created item is not
immediately available on cirrus. And even when it is available to
cirrus, this still takes time to propagate to all database replicas.
This becomes an issue during automated browser testing on systems that
are similar to production.

Specifically, the items for Lexeme language and lexical category have to
be known to cirrus before they can be added via the repo ui.

This implementation tries to renter the values when it isn't found until
it becomes known and then progresses as usual.

ToDo:
- find a good place for these helper methods

Bug: T222449
Change-Id: I7575f2ad34beb0fee1957fab9de6d1ab23dc024b