Add an empty line between `<?php` and the namespace
It appears this is the standard all code follows, with these few
exceptions.

The removal of the trait is unrelated, but I figured it's not worth
having it's own patch. If you disagree, please feel free to split this
patch. The reason for the removal is that it is already included in the
MediaWikiTestCase base class.

Change-Id: I0da87db3642116e7a30c02d23b1a4fa0285b135f