Wait for value to fix flaky test
This test started to become flaky in Gate (and CI?). It is unclear why
it starts to become flaky now, maybe performance has degraded so much?

Waiting for the value to be there seems to fix test.

For some unfathomable reason, waiting for that value inside of
`LexemePage.getNthStatementDataFromMainStatementGroup` doesn't work.

Bug: T225017
Change-Id: I802afda4ee7dca36d8db8f0b3c696ca7fd6d9b63
Depends-On: I91951ea03586ef4f11bb9d4676c2df6673f13380