Update phan, cleanup config
Explicitly require phan in composer, instead of letting the runner
auto-pick a version.
Remove useless copypasta in the config from wikibase, which adds a lot
of duplication with core config, and also references to non-existing
directories. Also try to remove all suppressions.
This is the standard way for extensions to configurate phan.

Bug: T226244
Bug: T226254
Bug: T227270
Bug: T224775
Change-Id: Ie09bf78cdc17f6a899802747685a72425fd46f94