Make mw.wikibase.lexeme.entity.lexeme inherit mw.wikibase.entity
We build a custom metatable that falls back to the methods of the
originally returned entity, so that you can also call generic methods
like getBestStatements() on the returned lexeme entity. This also means
we no longer need to implement getId() ourselves.

The implementation is a bit tricky because mw.wikibase.entity doesn’t
return its methodtable, so we actually have to call its create() method
and modify the returned table and its metatable, instead of just
assigning its methodtable as the __index of our methodtable’s metatable
and calling it a day.

Bug: T235901
Change-Id: I818f76d1514ce206b47dc844cef0fb5966af9b62