Autofocus inputs in glosses widget
Focus the first input when the user starts editing, and the first input
of the last table row (i. e. the language of the new gloss) when adding
a new gloss. (querySelector() and :last-child are both supported since
IE9, so that shouldn’t be a problem.)

One slightly ugly thing about this is the hidden dependency between
GlossWidget’s add function and its template (defined in PHP), due to the
tr in the selector. I think that’s acceptable, though. (We could also
use the .wikibase-lexeme-sense-gloss class instead.)

The focusing code is extracted into a utility function, focusElement,
because I’ll use it in other widgets as well.

Bug: T203461
Change-Id: I8519961314d7c42aeacab7f7c7c9e93a5c82dee1