Autofocus inputs in lemmas widget
Focus the first input when the user starts editing, and the first input
of the second-to-last list item (i. e. the lemma text of the new lemma:
the last list item is the “add” button) when adding a new lemma. (The
:nth-last-child pseudo-class is supported since IE9.)

As in the parent change (I8519961314), the dependency on the DOM layout,
which is split between the template and the corresponding JS code, is
slightly ugly, but acceptable in my opinion.

Note that, in Firefox, the very limited width of the lemma input can
give the appearance that the focus was not moved when editing an
existing lemma, because the cursor starts out at the end of the text but
Firefox doesn’t scroll the contents so the cursor is visible (see bug
1632896 [1]). Move the cursor a bit with the arrow keys to confirm that
the input is indeed focused.

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=1632896

Bug: T201588
Change-Id: I741b91079345d0fda4543ccc623c1c770443e71b