Form Representations: labels and input fields not connected
The input and label of Lexeme forms were not connected using label or
the 'for' attribute. Added test for confirming the existance of a label
with corresponding for attribute.

Awaiting refactoring of formIndex not being incremented each component
are currently using a custom uid solution to connect the label with the
corresponding input. See https://phabricator.wikimedia.org/T252516 for
more information.

Bug: T198702
Change-Id: Ib2a96f4498b4ccbde2a4864586abd5efc67f7676