Form template: Inconsistencies between production and test
During some work on the template for the form representations I
discovered inconsistencies between production and test, which might be
due to the test template being manually updated when changes are made
to the production one. I propose going forward we simply copy and paste
the template described in FormsView.php Line 165 to avoid this happening.

Change-Id: Icbe48606dd80921454744b55b232bba7d3ac2b50