Remove usage of EntityInfo
This depends on a change in Wikibase that allows methods in ViewFactory
to have null passed to them instead of an EntityInfo object.

In the view factory callback:
A default of null is needed and the typehint is removed so that Wikibase
can stop passing anything here.
Then in a separate change this parameter can be entirely removed.

Bug: T254283
Depends-On: I3801b4d1f4a876a820c28f4541ec0802d886b7eb
Change-Id: I727df0153b11c20c17ca26179a619b2b5728fa05