Skip Scribunto tests properly if client is not enabled
Right now, if you disable client, the tests explode with class not found
error. This fixes the problem in a similar manner client itself skips
the tests if Scribunto extension is not loaded.

Also merging these two conditions (Scribunto being enabled and
WikibaseClient) to one, so the error is more accurate. Currently, it
doesn't run any tests without saying why if Scribunto is not installed.
This patch fixes that.

Change-Id: Ibd4844352271769b9951798f3f1b707c9bd1d110