Disable exception logger for a test
If AbuseFilter is enabled, this test causes an exception to be logged,
which is supposed to make CI fail (the error log should be empty). The
correct solution for that isn’t clear (to be investigated in T255566),
but disabling exception logging for this test is good enough for now.

Bug: T249838
Change-Id: I718971c259ba0bb3dfca4fea329d442b23090d3b