Fix WikibaseLexemeExtensionRegistrationTest
The part of the test that runs when Repo is not enabled is failing on my
system: the message key is now paramvalidator-badvalue-enumnotmulti, not
apierror-unrecognizedvalue. I think this corresponds to the breaking API
change announced in February 2020 [1] – I can only assume that this part
of the test never failed in CI because the Repo is always enabled there.

[1]: https://lists.wikimedia.org/pipermail/mediawiki-api-announce/2020-February/000151.html

Change-Id: Iab3b048b12d9f94c9d84df80a5bb3464877736f9