Add toArrayUnordered() to SenseSet and FormSet
SenseSet::toArray() and FormSet::toArray() sort the list of Senses or
Forms before returning it, but sometimes we don’t care about that order:
specifically, when counting their statements (summed up into one number
anyways) or when extracting referenced entity IDs (order irrelevant).
Add a new method, toArrayUnordered(), which can be used in these cases.

(Note that I did not change StorageLexemeSerializer to use the new
methods. It would probably fine to do that, and write Senses and Forms
in whatever order they’re currently in – probably the order in which
they were stored in the previous revision – but I don’t feel confident
enough to do that.)

Change-Id: I52c6409f37d78b9f0971474412aea0926f6ed05e