LexemeSpecialEntityDataTest: construct via factory
This is not a unit tests but almost an edge2edge kind of situation so
the more realisitic the better, I think.

Ie41f287dc060a02a74258b35b23efbf7b937796f changes how the special page
is created and will benefit from this. It's a shame that the special
page is created to generate the link in the first place but as we are in
a SpecialPageTestBase the same method is used to generate the instance
to run the faux request test against.

Depends-On: Ibbe6ee9f547777fa62e4ef5f7998ba8281882e84
Change-Id: I245baa057db1695b35aed3dcc44227bda13e534d