Register all QUnit tests
When "packageFiles" is used, the first file is the “entry point” of the
module and other files are only loaded if they’re pulled in, directly or
transitively, through require() calls from that entry point. The usual
practice for QUnitTestModule seems to be to either use "scripts" instead
of "packageFiles" or to add a separate index file, and since some of our
tests seem to depend on packageFiles features, add the index file.

Generated from the extension.json list with the following command:

    jq -r '.QUnitTestModule.packageFiles | .[]' extension.json \
        | sed -n '/^tests\/qunit\// s//.\//p' \
	| tail -n+2 \
	| sed "s/\(.*\)/require( '\\1' );/" \
	>| tests/qunit/index.js

I also noticed that one test file was missing in the extension.json
list, so that’s added as well.

And one test in FormChanger had been broken by change I659c08783a
(commit c6e8e5c9ee) – the form callback parameter was renamed to
saveForm to placate some eslint rule forbidding variable shadowing, but
one assertion continued to refer to form rather than saveForm.

Bug: T264207
Change-Id: I4b9656a5a4f7b86ed8ac0d0ed6ac70099697977f