WikibaseLexeme: add test comparing Vue versions with MW
This test comapares the Vue and Vuex versions listed
on our package.json with the Mediawiki version.
Since we don't have Mediawiki access in CI this test uses
the npm package require-from-url in order to load the
vue file, so the version numbers may be extracted
and inspected.
In this case the test is located under mwlibs-testing
directory in order to allow it its own eslint rules
and avoid clashes with the main eslintrc.json

Bug: T259468
Change-Id: I7cd23af9bd169d0ce49ac4999a655c6f03d2d5bc