Don’t trim strings in JS
This effectively reverts the Vue and JS parts of change I711dc2b169
(commit 301cfe6f7e) – since all JS edits go through the API, it’s
sufficient to trim strings in the API, we don’t also need to do it in
JS, and especially not on every keystroke.

This does have the problem that, after a save, the UI sometimes still
thinks that a string has some leading or trailing spaces – the trimmed
version actually saved by the API doesn’t become visible until the page
is reloaded. But I think that can be fixed separately, and doesn’t need
to block this change.

Bug: T250550
Change-Id: I7af34273fb5341ca36d464458064074a61f118ef