Fix error handling in LexemeHeader
RepoApi.editEntity uses errorformat: 'plaintext' since change
Icf491206c1, which means that the response contains not a single error
field, but a list of potentially several errors. Handle this by taking
the first error of the list. Also, the error won’t have an info field;
depending on the formatversion, it’ll have either text or '*', so handle
both in displayError().

Also added the case of an error list in the relevant test.

Bug: T265783
Change-Id: Iba0721968ca0cc495fe720788b6b364aabcea4c6
Co-Authored-By: Noa Rave <noa.rave@wikimedia.de>